requirements.txt
streamlit>=1.35.0
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
pydantic>=2.0.0
pypdf>=4.0.0
python-docx>=1.1.0
python-dotenv>=1.0.0
